+++ MoviesUP/Source/Controller/Modules/Movies Detail/MovieDetailViewController.swift
@@ -62,41 +62,3 @@ class MovieDetailViewController: UIViewController {
         self.timeAndCategoriesLabel?.text =  movieDetail.timeAndCategories
     }
 }
-
-extension MovieDetailViewController: UIGestureRecognizerDelegate {
-    
-    func setInteractivePopGesture() {
-        navigationController?.interactivePopGestureRecognizer?.delegate = self
-    }
-    
-    func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
-        if let navigationController = navigationController, navigationController.viewControllers.count > 1 {
-            return true
-        }
-        return false
-    }
-}
-
-extension MovieDetailViewController {
-    func getMovieDetail() {
-        guard let movieID = movieID else { return }
-
-        ServiceManager().getMovieDetail(movieId: movieID) { (detail, error) in
-            guard let detail = detail else {
-                return
-            }
-            self.movieDetail = detail
-        }
-    }
-    
-    func getCredits() {
-        guard let movieID = movieID else { return }
-
-        ServiceManager().getCredits(movieId: movieID) { (cast, error) in
-            guard let cast = cast else {
-                return
-            }
-            self.cast = cast
-        }
-    }
-}
